Move auth cache into its own package
Was getting an import cycle in test mocks as the mocks imported auth which
then imported the mocks in the test file
