+++ auth/cache/cache.go
@@ -1,6 +1,7 @@
-package auth
+package cache
 
 import (
+	"github.com/kbase/blobstore/auth"
 	gcache "github.com/patrickmn/go-cache"
 )
 
@@ -9,17 +10,17 @@ import (
 // Cache caches auth service data from a provider.
 type Cache struct {
 	cache *gcache.Cache
-	prov  Provider
+	prov  auth.Provider
 }
 
 // NewCache creates a new auth cache.
-func NewCache(prov Provider) *Cache {
+func NewCache(prov auth.Provider) *Cache {
 	return &Cache{nil, prov}
 }
 
 // GetUser gets a user given a token.
 // Returns InvalidToken error.
-func (c *Cache) GetUser(token string) (*User, error) {
+func (c *Cache) GetUser(token string) (*auth.User, error) {
 	u, _, _, err := c.prov.GetUser(token)
 	if err != nil {
 		return nil, err
